Move storybook templates function to utils folder
